feat: organize components and pages
