+++ phone-test/src/pages/CallDetails/index.ts
@@ -1,0 +1,1 @@
+export { CallDetailsPage as default } from './CallDetails';
