+++ phone-test/src/pages/CallList/index.ts
@@ -1,0 +1,1 @@
+export { CallsListPage as default } from './CallsList';
